requirements.txt
numpy>=1.26,<3.0
tensorflow>=2.15,<2.22
streamlit>=1.35,<2.0
pillow>=10.0
matplotlib>=3.8
scipy>=1.11
scikit-learn>=1.4
seaborn>=0.13
kaggle>=1.6

